Start of mat3 refactor
